% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileList.R
\name{FileList}
\alias{FileList}
\title{Create file list}
\usage{
FileList(path, tmstamp)
}
\arguments{
\item{path}{character: the path to the directories containing folders of data}

\item{tmstamp}{character: date field to return, which should be the date for
which the data was taken; either "mtime" (date
modified), "ctime" (last status change), or "atime" (last access) -- this
value should usually be "mtime"}
}
\value{
csv file of file list in the input directory
}
\description{
\code{FileList} creates file list in any directory
}
